Adjust click handling to fix zoom out
Original code used isPopUpTrigger() to determine if the mouse press
was RMB (as on windows, RMB triggers the pop-up menu). However, this
isn't defined properly on other platforms, so it was only possible to
zoom in, never out. Replace with a check for !BUTTON1; any button
other than 1 zooms out.