WadCMainFrame,WadCPrefs: begin decoupling
Begin decoupling MainFrame from the rest of WadC, in order to
provide alternative UIs such as a CLI.

WadCMainFrame is an interface that captures the usage of MainFrame
in WadC at the moment (with some adjustments to both).

WadCPrefs is a struct to collect the fields that are read or written
to the prefs file.