Tweak RNG behaviour to print out seed
Adjust the handling of the RNG for the Choice operator so we can print
out the seed that is used (based on gettimeofday) when the script is
first evaluated. This means if someone generates something really cool
they could save the seed value that generated that cool thing.

Fix a bug whereby re-running a script that did some choice stuff w/o
setting a fixed seed value, but later set the fixed seed value, meant
the RNG was fixed (to that value) for subsequent invocations.

Add a test file to demonstrate the above.