Fix up the regression test suite
Move the Makefile to the top-level. We want to check against examples/
and this avoids needing hacky path stuff or handling "..". Use the
hidden "-nosrc" option to not embed the source, which would guarantee
the sums changing anyway.

Update sha1sums for the current state of WadC, the .wl files and my
local Java version (Oracle JRE build 1.8.0_45-b14)