refresh sha1sums + add new test
NOTE: this test was uncommitted earlier, but the checksum snuck into
the sha1sums file, and it's changed, but I don't think it should
have. Having said that I don't know what version of the WAD was the
one that was originally checksummed and because it wasn't committed
I can't easily find out.