built-in forceangle, for thing angles
New builtin forceangle(x) which forces the angle value used
for all subsequent things.

Normally WadC uses a (derivation of) the cursor orientation,
but for poly objects, Raven overrode the thing angle for the
polyobj ID field, so we need to be able to set things exactly.

I might rethink this before releasing WadC with it in; also,
there's currently no way to make it go back to tracking the
cursor orientation after you've called forceangle(x).