WIP don't use linedef tag value as arg[0] in hexen
This confused me. The old behaviour was to write the first argument
for setlinetypehexen into the Linedef 'tag' field, even though
Hexen doesn't have tags. The 'args' field was 5-values long, but
the first value was completely ignored (although setlinetypehexen
wrote to it, it was never written out to WADs).

The ZDoom examples freely mixed setlinetypehexen and setlinetype
which is why they worked; rework them to choose the hexen scheme
only. (I guess it would be good to be able to mix and match. Hm.)

Further thought required here.

TODO: remove the sha1sums change from this commit