built-in thingangle, for thing angles
New builtin thingangle(x) to create a new thing with a specific facing of x.

Normally WadC uses a (derivation of) the cursor orientation, but for poly-
objects, Raven overrode the thing angle for the polyobj ID field, so we need to
be able to set it exactly.

Adjust makething() to accept the angle as a parameter and adjust the existing
thing builtin accordingly.

Document thingangle and put the angle constants for 8 compass point angles in
standard.h.

Add a test tests/angles.wl.