Replace forceangle built-in with thingangle
Remove forceangle and the manual_angle boolean and instead add a
new unary builtin thingangle() which accepts an angle value as
the argument. This value is used for the creation of that thing
only.

Adjust makething() to accept the angle as a parameter and adjust
the existing thing builtin accordingly.

Update polyobj.wl to use thingangle().

regression test suite shows that the WAD output is identical.