pipes: improve the lift
adjust the 'slime2' settings so slime is 16px above floor level and
move the setting of it prior to slimequad. (I will later on want at
least one of the closed branches from slimequad to use slime1's
settings so further work here needed)

lift is 24px tall so above water level and doesn't need tagging.

use a boom generalized linedef for the lift action so it returns to
the 24px position and not lowest neighbour (below water level).
add a nearest-neighbour to the dummy sector.

I'm amazed this works.