beta up3: layout experiments for pipes
How many times will I restart this from scratch? This is a from-scratch
restart of the layout of "pipes", and some corresponding changes to
pipes to support that (which might have broken usepipes...)

Two circles of pipes that intersect twice like a figure of eight.

slimequad reworked so you can pass a slime object to it and the
unneeded tag argument is gone

slimelift adjusted so the lift is large enough to fill up the space
and you can't get below it if its up (although you can still activate
it, that needs tweaking). Lift start height configurable