Update water to use a patch in doom 1 & 2
To define a COLORMAP for use in underwater regions, Boom's 242
linedef overloads the mid texture of the sidedef for the control
line. If you try to play such a map in vanilla, it will crash
due to the missing texture.

water.h defines a compatibility texture with a name matching the
COLORMAP to avoid this vanilla crash.

Change the patch used for this dummy texture to SW12_1, which is
in both Doom and Doom 2.