move toggle{things,vertices} to prefs; deprecate
Mark a whole bunch of built-ins as deprecated. These are the preferences
built-ins, which I am going to remove from the language.

Pull togglethings and togglevertices up out of WadRun and make them
members of the WadCPrefs object.

(A little bit of iteration tidy-up on the lines I've modified for the
above)

This changes their behaviour slightly: because the value persists between
runs, it will not default to true at the start of every run anymore.

remove use of toggle{things,verts} in examples.