WIP: use picocli for WadCCLI command line handling
picocli is licensed under the Apache License 2.0. TODO need to double
check the implications of using and bundling it.

This improves the existing command-line option handling and lets us
expand to add --seed for specifying the initial RNG seed. This will
eventually let me move the Makefile-driven sha1-checksum stuff to
force a particular seed for reproducibility, and we can have the
examples themselves be randomised.

I also reformatted the Java file brace syntax since I rewrote so
much.

Adjust the maven-assembly-plugin configuration to build two
assemblies: first, a jar-with-dependencies, including picocli.
Second, the zip assembly, including the jar-with-dependencies.
Adjust that assembly configuration to ignore the "normal" jar.

Remove maven-jar-plugin configuration; we don't want the
dependency-less JAR (TODO: or do we? Debian probably does.)
Although it gets pulled in and executed anyway as a dependency
of maven-assembly-plugin.