update sha1sums for fixed-RNG results
Adjust all the examples which have a random element so that repeated
re-runs of the example in the WadC GUI will give different results.
It's nice to repeatedly re-run to see how varied the output can be.

Force the initial RNG seed to a fixed number when generating the
test WAD files via Makefile. This ensures the output is reproducible.

Recalculate all the sha1sums and save their values.