TuneablePanel: clean up when refreshing the UI
Move adding seedPanel to the RandomListener list out of a constructor
(only called once) and into the method that is repeatedly called to
populate the UI

When we remove a seedPanel, unregister it as a RandomListener.

This addresses an issue where the RNG UI stops being functional after
you load a new WadL file.